Workaround for compiler complaining about Affine.m(int) being ambiguous with m(const double*)

+++ lax/interfaces/captioninterface.cc
@@ -708,7 +708,7 @@ void CaptionData::dump_out(FILE *f,int indent,int what,Laxkit::DumpContext *cont
 
 
 	fprintf(f,"%smatrix %.10g %.10g %.10g %.10g %.10g %.10g\n",spc,
-				m(0),m(1),m(2),m(3),m(4),m(5));
+				m((int)0),m(1),m(2),m(3),m(4),m(5));
 
 	fprintf(f,"%sfont\n",spc);
 	Attribute att;
@@ -755,7 +755,7 @@ Laxkit::Attribute *CaptionData::dump_out_atts(Attribute *att,int what,DumpContex
 	}
 
 	char scratch[200];
-	sprintf(scratch, "%.10g %.10g %.10g %.10g %.10g %.10g", m(0),m(1),m(2),m(3),m(4),m(5));
+	sprintf(scratch, "%.10g %.10g %.10g %.10g %.10g %.10g", m((int)0),m(1),m(2),m(3),m(4),m(5));
 	att->push("matrix", scratch);
