+++ lax/interfaces/roundedrectinterface.cc
@@ -320,7 +320,7 @@ Attribute *RoundedRectData::dump_out_atts(Attribute *att,int what,Laxkit::DumpCo
 
 	att->push("id", Id());
 	att->pushStr("matrix",-1, "%.10g %.10g %.10g %.10g %.10g %.10g",
-			m(0),m(1),m(2),m(3),m(4),m(5));
+			m((int)0),m(1),m(2),m(3),m(4),m(5));
 
 	if (is_square) att->push("square", "yes");
